Add files via upload
During the first year of my Bachelor we got the assignment to choice any one art piece at the university and make a game using that. The background painting is the art piece we went with as it reminded us of a ghost town.
In this small project, you are a ghost Lightbringer and need to turn on the lights of the town.